Fix a bad typeof comparison
This was a delightfully subtle bug caught by DeepScan, with the effect
that setting CalendarEvent#start for the first time would have mangled
CalendarEvent#recurrenceOverrides if it was already set. (It’d end up as
an object with “null” as its only key, and an arbitrary one of the
recurrenceOverride values as its value.) In practice I *think* it was
pretty harmless.